+++ test/jest/test.js
@@ -130,7 +130,6 @@ if (TEST_CTRL.PARSE_CONFIG) {
       'concat': {},
       'resource': {},
       'plugins': [
-        'yyl-flexlayout',
         'yyl-flexlayout'
       ],
       'webpackConfigPath': `${configDir}/webpack.config.js`,
@@ -153,7 +152,81 @@ if (TEST_CTRL.PARSE_CONFIG) {
         '@': `${configDir}/src`,
         '~@': `${configDir}/src/components`
       },
-      'resolveModule': `${yh.vars.SERVER_PLUGIN_PATH}/webpack-vue3/node_modules`
+      'resolveModule': `${yh.vars.SERVER_PLUGIN_PATH}/webpack-vue3/1/node_modules`
+    };
+
+    expectResult.concat[`${configDir}/dist/project/1/mobile/js/vendors.js`] = [
+      `${configDir}/src/js/lib/a.js`,
+      `${configDir}/src/js/lib/b.js`
+    ];
+
+    expectResult.resource[`${configDir}/src/pc/svga`] = `${configDir}/dist/project/1/mobile/tpl`;
+    expect(r).toEqual(expectResult);
+  });
+
+  test('yh.parseConfig(configPath, iEnv, returnKeys): object no plugins', async () => {
+    const configDir = util.path.join(__dirname, '../case/case-parse-config-noplugins');
+    const configPath = path.join(configDir, 'yyl.config.js');
+    yh.setVars({ PROJECT_PATH: configDir });
+    const r = await yh.parseConfig(configPath);
+    const expectResult = {
+      'workflow': 'webpack-vue2',
+      'name': '1',
+      'version': '3.4.10',
+      'platform': 'mobile',
+      'proxy': {
+        'port': 8887,
+        'localRemote': {
+          'http://web.yy.com/': 'http://127.0.0.1:5000/',
+          'http://www.yy.com/web/1': 'http://127.0.0.1:5000/project/1/mobile/html',
+          'http://www.yy.com/api/mock': 'http://127.0.0.1:5000/api/mock'
+        },
+        'homePage': 'http://www.yy.com/web/1/'
+      },
+      'localserver': {
+        'root': './dist',
+        'port': 5000
+      },
+      'dest': {
+        'basePath': '/project/1/mobile',
+        'jsPath': 'js',
+        'jslibPath': 'js/lib',
+        'cssPath': 'css',
+        'htmlPath': 'html',
+        'imagesPath': 'images',
+        'tplPath': 'tpl',
+        'revPath': 'assets'
+      },
+      'commit': {
+        'type': 'gitlab-ci',
+        'revAddr': 'http://web.yystatic.com/project/1/mobile/assets/rev-manifest.json',
+        'hostname': '//web.yystatic.com',
+        'staticHost': '//web.yystatic.com',
+        'mainHost': '//www.yy.com/web'
+      },
+      'concat': {},
+      'resource': {},
+      'plugins': [],
+      'webpackConfigPath': `${configDir}/webpack.config.js`,
+      'alias': {
+        'root': `${configDir}/dist/project/1/mobile`,
+        'revRoot': `${configDir}/dist/project/1/mobile`,
+        'destRoot': `${configDir}/dist`,
+        'srcRoot': `${configDir}/src`,
+        'dirname': `${configDir}`,
+        'commons': util.path.join(configDir, '../commons'),
+        'globalcomponents': util.path.join(configDir, '../commons/components'),
+        'globallib': util.path.join(configDir, '../commons/lib'),
+        'jsDest': `${configDir}/dist/project/1/mobile/js`,
+        'jslibDest': `${configDir}/dist/project/1/mobile/js/lib`,
+        'htmlDest': `${configDir}/dist/project/1/mobile/html`,
+        'cssDest': `${configDir}/dist/project/1/mobile/css`,
+        'imagesDest': `${configDir}/dist/project/1/mobile/images`,
+        'revDest': `${configDir}/dist/project/1/mobile/assets`,
+        'tplDest': `${configDir}/dist/project/1/mobile/tpl`,
+        '@': `${configDir}/src`,
+        '~@': `${configDir}/src/components`
+      }
     };
 
     expectResult.concat[`${configDir}/dist/project/1/mobile/js/vendors.js`] = [
@@ -179,7 +252,7 @@ if (TEST_CTRL.OPTIMIZE) {
     const config = await yh.parseConfig(configPath, iEnv);
     const revPath = path.join(config.alias.revDest, yh.optimize.rev.filename);
 
-    const serverPluginPath = path.join(yh.vars.SERVER_PLUGIN_PATH, config.workflow);
+    const serverPluginPath = path.join(yh.vars.SERVER_PLUGIN_PATH, config.workflow, config.name);
     await extFs.removeFiles(serverPluginPath);
 
     // 检查结果